Add versioning system
Now .jar-file will look like Atom-<branch>-<commit_hash> for usual
commits and Atom-<tag> for commits with a tag. Also, there is one more
part of the .jar-file name: -dirty for builds, which have been made from
staging sources. And 2 old endings: -dev and -server. For using as a
library and for using in a production accordingly.