Fix CraftBukkit packages(#9)
Some plugins get server version by extracting it from the CraftBukkit
path. Thanks to this mapping line it is possible to do this.
Known problem: the new package is not a replacement for
org/bukkit/craftbukkit, it is just created in a new directory
org/bukkit/craftbukkit/v1_12_R1. The only drawback of such behavior is
increased .jar-file size by 0.7 Mb.