Fix TileEntity snapshot creation
When someone tries to get mod block's state the CraftBlockEntityState is
created. To create that CraftBlockEntityState CraftBukkit makes a
TileEntity snapshot by creation another one TileEntity, but without any
world setting. So, there's always null world in such snapshot. As a
result NullPointerException is thrown while accessing to this world.

Closes #75.