Implement BlockPlaceEvent and BlockSnapshot to BlockState conversion
Forge uses BlockSnapshot while CraftBukkit uses BlockState. Both these
classes are almost the same, so we can use Forge storages for working
with them and convert BlockSnapshot to CraftBukkit representation named
CraftBlockState in the last moment, before passing it to the public part
of BukkitAPI.
This commit is mostly proof of concept and Atom cannot be used at this
state, as there should be some extra code for StructureGrowEvent, test
coverage and BlockState/BlockSnapshot storages deduplication.
May be broken.

Closes #87.