Fix player burning on respawn after death in lava
CraftBukkit calls EntityCombustByBlockEvent when `fire` counter is <= 0.
This statement is true when player jumps into lava for the first time.
But, it is also true, when player has died in lava, so the event is
fired event the player is dead, which is incorrect. In addition to event
firing CraftBukkit also sets fire to the player, which causes player
burning after respawn and the player can see himself burning for a short
time. Thus, we now check if player is alive and only then fire the event
and set fire to the player.

Closes #112.